fix increment user version after update
Took 42 seconds

+++ src/main/java/com/github/gribanoveu/auth/entities/tables/User.java
@@ -22,7 +22,7 @@ public class User {
     private Long id;
 
     @Version
-    private int version;
+    private Long version;
 
     @Column(name = "email")
     private String email;
